Relicense to Apache 2.0
I am on the only non-trivial copyright holder of this repository
Now that our scratch-vm has relicensed, now's a great time to make the same change to the entirety of the packager
Following the full LGPL in the context of the packager is unnecessarily difficult. Apache 2.0 does not have these problems.