Remove service worker
We have never properly embraced service workers.
Offline support is an afterthought that hardly works, and our site is not even installable as a PWA
We should instead properly advertise the standalone versions.
This has been the source of many bug reports over time, so let's just get rid of it.