Disable enforcePrivacy in packaged projects
The threat model of enforcePrivacy is so that people can open up sb3 files
in the editor and not worry about their webcam being streamed to an
attacker if the project doesn't use any unsandboxed custom extensions.

That threat model doesn't really make sense in the packager when we're
generating HTML files.